+++ SavingsGoalManager/src/main/java/com/brenner/budgetmanager/savingsgoals/SavingsGoalDepositAllocation.java
@@ -1,14 +1,16 @@
 package com.brenner.budgetmanager.savingsgoals;
 
+import java.math.BigDecimal;
+
 /**
  * A special data transfer object that encapsulates the allocation of a part of a deposit to a specific goal.
  */
 public class SavingsGoalDepositAllocation {
     Integer savingsGoalId;
     Long depositId;
-    Float allocationAmount;
+    BigDecimal allocationAmount;
     
-    public SavingsGoalDepositAllocation(Integer savingsGoalId, Long depositId, Float allocationAmount) {
+    public SavingsGoalDepositAllocation(Integer savingsGoalId, Long depositId, BigDecimal allocationAmount) {
         this.savingsGoalId = savingsGoalId;
         this.depositId = depositId;
         this.allocationAmount = allocationAmount;
@@ -30,11 +32,11 @@ public class SavingsGoalDepositAllocation {
         this.depositId = depositId;
     }
     
-    public Float getAllocationAmount() {
+    public BigDecimal getAllocationAmount() {
         return allocationAmount;
     }
     
-    public void setAllocationAmount(Float allocationAmount) {
+    public void setAllocationAmount(BigDecimal allocationAmount) {
         this.allocationAmount = allocationAmount;
     }
 }
